+++ common/src/main/java/earth/terrarium/botarium/common/energy/impl/InsertOnlyEnergyContainer.java
@@ -24,7 +24,7 @@ public class InsertOnlyEnergyContainer extends SimpleEnergyContainer {
 
     @Override
     public long internalExtract(long maxAmount, boolean simulate) {
-        long extracted = Mth.clamp(maxAmount, 0, getStoredEnergy());
+        long extracted = (long) Mth.clamp(maxAmount, 0, getStoredEnergy());
         if(simulate) return extracted;
         this.setEnergy(getStoredEnergy() - extracted);
         return extracted;
